Fix up RSpec/DescribedClass issue with ndjson spec
